Day 5 part 1,2 : With some nasty brute force garbage.

+++ src/main/java/com/ilaird/aoc2023/aoc/Solver.java
@@ -1,5 +1,5 @@
 package com.ilaird.aoc2023.aoc;
 
 public interface Solver {
-    int solve() throws SolverError;
+    long solve() throws SolverError;
 }
